fresh 1.8 への移行をロールバック
