+++ .github/workflows/deploy.yml
@@ -7,29 +7,20 @@ on:
   workflow_dispatch:
 
 permissions:
-  contents: read   # allow GitHub to read repository contents
-  id-token: write  # required for OIDC authentication with Deno Deploy
+  contents: read # allow GitHub to read repository contents
+  id-token: write # required for authentication with Deno Deploy
 
 jobs:
   deploy:
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Setup Deno
-        uses: denoland/setup-deno@v2
+      - uses: actions/checkout@v3
+      - uses: denoland/setup-deno@v1
         with:
           deno-version: v2.x
-
-      - name: Format code and run checks
-        run: |
-          deno fmt
-          deno task check
-
-      - name: Deploy to Deno Deploy
-        uses: denoland/deployctl@v1
+      - run: deno task check
+      - uses: denoland/deployctl@v1
         with:
-          project: namnam-cosense-exp-47  # Deno Deploy project name
-          root: "."                       # Root directory for deployment bundle
-          entrypoint: main.ts              # Application entrypoint file
+          project: namnam-cosense-exp-47
+          entrypoint: server.ts
+          token: ${{ secrets.DENO_DEPLOY_ACCESS_TOKEN }}
