+++ deno.json
@@ -1,10 +1,9 @@
 {
-  "importMap": "./import_map.json",
   "tasks": {
-    "start": "deno run -A main.ts",
+    "start": "deno run --allow-net --allow-env server.ts",
     "fmt": "deno fmt",
     "lint": "deno lint",
     "check": "deno lint --ignore=docs && deno fmt --check --ignore=docs",
-    "deploy": "deployctl deploy --project=namnam-cosense-exp-47 main.ts"
+    "deploy": "deployctl deploy --project=namnam-cosense-exp-47 server.ts"
   }
 }
