--- fresh.config.ts
@@ -1,3 +1,0 @@
-import { defineConfig } from "$fresh/server.ts";
-
-export default defineConfig({});
