--- import_map.json
@@ -1,6 +1,0 @@
-{
-  "imports": {
-    "$fresh/": "https://deno.land/x/fresh@1.7.3/",
-    "$std/": "https://deno.land/std@0.203.0/"
-  }
-}
