--- main.ts
@@ -1,4 +1,0 @@
-import { start } from "$fresh/server.ts";
-import manifest from "./fresh.gen.ts";
-
-await start(manifest);
